# web_app/docs/swagger.yaml
basePath: /
definitions:
  models.Comment:
    properties:
      content:
        description: 评论内容
        type: string
      created_at:
        description: 创建时间
        type: string
      id:
        description: 评论ID（JSON序列化为字符串以避免JavaScript精度丢失）
        example: "0"
        type: string
      parent_id:
        description: 父评论ID（用于回复，可为空）
        example: "0"
        type: string
      topic_id:
        description: 话题ID
        example: "0"
        type: string
      updated_at:
        description: 更新时间
        type: string
      user_id:
        description: 用户ID
        example: "0"
        type: string
      username:
        description: 用户名（从users表JOIN）
        type: string
    type: object
  models.CommentListResponse:
    properties:
      comments:
        description: 评论列表
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      has_more:
        description: 是否有下一页
        type: boolean
      page:
        description: 当前页
        type: integer
      page_size:
        description: 每页数量
        type: integer
      total:
        description: 总数
        type: integer
      total_pages:
        description: 总页数
        type: integer
    type: object
  models.CreateCommentRequest:
    properties:
      content:
        description: 评论内容：1-1000个字符
        maxLength: 1000
        minLength: 1
        type: string
      parent_id:
        description: 父评论ID（可选，用于回复）
        type: integer
    required:
    - content
    type: object
  models.CreateTopicRequest:
    properties:
      category:
        description: 分类
        enum:
        - tech
        - design
        - discuss
        - share
        - product
        type: string
      content:
        description: 内容：至少10个字符
        minLength: 10
        type: string
      title:
        description: 标题：5-100个字符
        maxLength: 100
        minLength: 5
        type: string
    required:
    - category
    - content
    - title
    type: object
  models.LoginRequest:
    properties:
      password:
        description: 密码
        type: string
      username:
        description: 用户名
        type: string
    required:
    - password
    - username
    type: object
  models.LoginResponse:
    properties:
      token:
        description: JWT token
        type: string
      user_id:
        description: 用户ID（JSON序列化为字符串）
        example: "0"
        type: string
      username:
        description: 用户名
        type: string
    type: object
  models.RegisterRequest:
    properties:
      email:
        description: 邮箱：必须是有效的邮箱格式
        type: string
      password:
        description: 密码：至少6个字符
        minLength: 6
        type: string
      username:
        description: 用户名：4-20个字符
        maxLength: 20
        minLength: 4
        type: string
    required:
    - email
    - password
    - username
    type: object
  models.Response:
    properties:
      code:
        description: 状态码：0=成功，其他=失败
        type: integer
      data:
        description: 响应数据（可选）
      message:
        description: 提示信息
        type: string
    type: object
  models.SearchResponse:
    properties:
      has_more:
        description: 是否有下一页
        type: boolean
      page:
        description: 当前页
        type: integer
      page_size:
        description: 每页数量
        type: integer
      took:
        description: 搜索耗时(毫秒)
        type: integer
      topics:
        description: 话题列表
        items:
          $ref: '#/definitions/models.Topic'
        type: array
      total:
        description: 总数
        type: integer
      total_pages:
        description: 总页数
        type: integer
    type: object
  models.Topic:
    properties:
      category:
        description: 分类（tech/design/discuss/share/product）
        type: string
      comment_count:
        description: 评论数
        type: integer
      content:
        description: 话题内容
        type: string
      created_at:
        description: 创建时间
        type: string
      dislike_count:
        description: 点踩数
        type: integer
      id:
        description: 话题ID（JSON序列化为字符串以避免JavaScript精度丢失）
        example: "0"
        type: string
      like_count:
        description: 点赞数
        type: integer
      title:
        description: 话题标题
        type: string
      updated_at:
        description: 更新时间
        type: string
      user_id:
        description: 发布者ID
        example: "0"
        type: string
      username:
        description: 发布者用户名（从users表JOIN）
        type: string
      view_count:
        description: 浏览数
        type: integer
    type: object
  models.TopicListResponse:
    properties:
      has_more:
        description: 是否有下一页
        type: boolean
      page:
        description: 当前页
        type: integer
      page_size:
        description: 每页数量
        type: integer
      topics:
        description: 话题列表
        items:
          $ref: '#/definitions/models.Topic'
        type: array
      total:
        description: 总数
        type: integer
      total_pages:
        description: 总页数
        type: integer
    type: object
  models.User:
    properties:
      created_at:
        description: 创建时间
        type: string
      email:
        description: 邮箱
        type: string
      id:
        description: 用户ID（JSON序列化为字符串以避免JavaScript精度丢失）
        example: "0"
        type: string
      updated_at:
        description: 更新时间
        type: string
      username:
        description: 用户名
        type: string
    type: object
host: localhost:8082
info:
  contact:
    name: DancingCircles
    url: https://github.com/DancingCircles/Bullbell
  description: 基于Go+Gin的论坛系统API，支持用户注册登录、话题发布、评论互动等功能
  title: Bullbell Forum API
  version: "1.0"
paths:
  /api/v1/comments/{id}:
    delete:
      description: 删除自己的评论
      parameters:
      - description: 评论ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 删除评论
      tags:
      - 评论
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: 用户登录获取token
      parameters:
      - description: 登录信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.LoginResponse'
              type: object
      summary: 用户登录
      tags:
      - 用户
  /api/v1/register:
    post:
      consumes:
      - application/json
      description: 创建新用户账号
      parameters:
      - description: 注册信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: 用户注册
      tags:
      - 用户
  /api/v1/search:
    get:
      description: 全文搜索话题，支持关键词、分类筛选和排序
      parameters:
      - description: 搜索关键词
        in: query
        name: keyword
        type: string
      - description: 分类筛选
        in: query
        name: category
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - default: created_at
        description: 排序方式：created_at/view_count/comment_count
        in: query
        name: sort_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.SearchResponse'
              type: object
      summary: 搜索话题
      tags:
      - 搜索
  /api/v1/search/hot:
    get:
      description: 获取指定分类的热门话题（按浏览量排序）
      parameters:
      - description: 分类
        in: query
        name: category
        required: true
        type: string
      - default: 10
        description: 数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Topic'
                  type: array
              type: object
      summary: 获取分类热门话题
      tags:
      - 搜索
  /api/v1/search/stats:
    get:
      description: 统计各分类的话题数量
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  additionalProperties:
                    format: int64
                    type: integer
                  type: object
              type: object
      summary: 获取分类统计
      tags:
      - 搜索
  /api/v1/search/suggest:
    get:
      description: 根据前缀提供搜索建议
      parameters:
      - description: 搜索前缀
        in: query
        name: prefix
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
      summary: 搜索建议
      tags:
      - 搜索
  /api/v1/topics:
    get:
      description: 分页获取话题列表，支持排序和筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - default: hot
        description: 排序方式：hot/new/like
        in: query
        name: sort
        type: string
      - description: 分类筛选
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.TopicListResponse'
              type: object
      summary: 获取话题列表
      tags:
      - 话题
    post:
      consumes:
      - application/json
      description: 发布新话题
      parameters:
      - description: 话题信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateTopicRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 创建话题
      tags:
      - 话题
  /api/v1/topics/{id}:
    get:
      description: 根据ID获取话题详细信息
      parameters:
      - description: 话题ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Topic'
              type: object
      summary: 获取话题详情
      tags:
      - 话题
  /api/v1/topics/{id}/comments:
    get:
      description: 分页获取话题的评论列表
      parameters:
      - description: 话题ID
        in: path
        name: id
        required: true
        type: integer
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.CommentListResponse'
              type: object
      summary: 获取话题评论列表
      tags:
      - 评论
    post:
      consumes:
      - application/json
      description: 对话题发表评论
      parameters:
      - description: 话题ID
        in: path
        name: id
        required: true
        type: integer
      - description: 评论信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 创建评论
      tags:
      - 评论
  /api/v1/topics/{id}/vote:
    post:
      consumes:
      - application/json
      description: 对话题进行点赞或点踩
      parameters:
      - description: 话题ID
        in: path
        name: id
        required: true
        type: integer
      - description: 投票类型：like/dislike
        in: query
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 给话题投票
      tags:
      - 话题
  /api/v1/user/info:
    get:
      description: 根据token获取用户信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取当前用户信息
      tags:
      - 用户
securityDefinitions:
  ApiKeyAuth:
    description: JWT token，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"